test/reftest/vendor/blink/fast/dom/DocumentType/next-element-sibling.html

<!DOCTYPE html>
<script src="../../../resources/wpt-test-pre.js"></script>
<script src="../../../resources/js-test-pre.js"></script>
<script>

description('This tests the NonDocumentTypeChildNode interface nextElementSibling property on a DocumentType node.');

var doctype = document.doctype;
shouldBeUndefined('doctype.nextElementSibling');

</script>
<script src="../../../resources/js-test-post.js"></script>
<script src="../../../resources/wpt-test-post.js"></script>
